requirements.txt
alembic==0.9.6
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
defusedxml==0.6.0
dominate==2.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
pbr==3.1.1
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
python3-openid==3.2.0
pytz==2017.2
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.1
